gpu: ion: Introduce ion_map_iommu_by_force()
This does the same thing as ion_map_iommu(), but if the handle that is attempting to be mapped is already mapped then it will be unmapped first, ensuring success. This should be used in areas where hard-to-debug memory leaks are expected and where the success of ion_map_iommu() is crucial.

Signed-off-by: Sultanxda <sultanxda@gmail.com>