msm_fb: Only wake the screen when power button is pressed
In Lollipop, mdp_on() is being called when the device is in suspend, causing the screen to power on during sleep. Ideally we could use early suspend to block rogue calls to mdp_on(), but using early suspend in msm_fb is freezing the device when it goes to sleep.

As a hack, only power on the display when mdp is asleep and power-button input is detected.

Signed-off-by: Sultanxda <sultanxda@gmail.com>